Made GameViewModel._stockWasteEmpty protected in order to make it false in YukonViewModel.init since all Cards start in Tableau. Hooked up YukonViewModel in MainActivity. No other code changes, created two new directories in .data: .pile and .pile.tableau. This is to better organize .data directory.
