+++ app/src/main/java/com/heyzeusv/solitaire/data/pile/Foundation.kt
@@ -1,5 +1,6 @@
-package com.heyzeusv.solitaire.data
+package com.heyzeusv.solitaire.data.pile
 
+import com.heyzeusv.solitaire.data.Card
 import com.heyzeusv.solitaire.util.Suits
 
 /**
