+++ app/src/main/java/com/heyzeusv/solitaire/data/pile/Waste.kt
@@ -1,4 +1,6 @@
-package com.heyzeusv.solitaire.data
+package com.heyzeusv.solitaire.data.pile
+
+import com.heyzeusv.solitaire.data.Card
 
 /**
  *  In Solitaire, Waste refers to the face up pile where cards drawn from [Stock] are placed. Only
