+++ app/src/main/java/com/heyzeusv/solitaire/ui/YukonViewModel.kt
@@ -36,5 +36,6 @@ class YukonViewModel @Inject constructor(
 
     init {
         resetAll(ResetOptions.NEW)
+        _stockWasteEmpty.value = true
     }
 }
