+++ app/src/test/java/com/heyzeusv/solitaire/data/FoundationTest.kt
@@ -1,5 +1,6 @@
 package com.heyzeusv.solitaire.data
 
+import com.heyzeusv.solitaire.data.pile.Foundation
 import com.heyzeusv.solitaire.util.Suits
 import com.heyzeusv.solitaire.util.TestCards
 import org.junit.Assert.assertEquals
