+++ app/src/test/java/com/heyzeusv/solitaire/data/StockTest.kt
@@ -1,5 +1,6 @@
 package com.heyzeusv.solitaire.data
 
+import com.heyzeusv.solitaire.data.pile.Stock
 import com.heyzeusv.solitaire.util.TestCards
 import org.junit.Assert.assertEquals
 import org.junit.Test
